Revert "Ver2"
