+++ Android App/app/src/main/java/com/example/app/Dashboard(Teacher).kt
@@ -57,7 +57,7 @@ import kotlinx.coroutines.delay
 @Composable
 fun DashboardTApp(navController: NavController) {
     var searchQuery by remember { mutableStateOf("") }
-    var filterOption by remember { mutableStateOf("Both") }
+    var filterOption by remember { mutableStateOf("Both") } // Filter: "Both", "Name", or "Tag"
     var items by remember { mutableStateOf<List<Map<String, Any>>>(emptyList()) }
     var errorMessage by remember { mutableStateOf<String?>(null) }
     var loading by remember { mutableStateOf(true) }
@@ -65,8 +65,6 @@ fun DashboardTApp(navController: NavController) {
     var refresh by remember { mutableStateOf(false) }
 
     val filterOptions = listOf("Both", "Name", "Tag")
-
-    UserTypeWatcher(navController, false, false)
 
     LaunchedEffect(Unit) {
         getItemsFromFirestore(
