+++ Android App/app/src/main/java/com/example/app/MainActivity.kt
@@ -23,7 +23,6 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.foundation.text.selection.TextSelectionColors
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.FilledTonalButton
@@ -31,12 +30,10 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.material3.OutlinedTextFieldDefaults
-import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -58,9 +55,6 @@ import com.google.firebase.auth.auth
 import androidx.navigation.compose.*
 import androidx.navigation.navArgument
 import com.google.firebase.firestore.DocumentReference
-import kotlinx.coroutines.coroutineScope
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
 
 class MainActivity : ComponentActivity() {
     @RequiresApi(Build.VERSION_CODES.O)
@@ -98,20 +92,8 @@ fun RootApp() {
         composable("login") {
             LoginApp(navController)
         }
-        composable("signup") {
-            SignUpScreen(navController)
-        }
         composable("viewItem") {
             ViewItemUi(navController)
-        }
-        composable("waiting") {
-            WaitingScreen(navController)
-        }
-        composable("management") {
-            ManagementScreen(navController)
-        }
-        composable("userManagement") {
-            UserManagementScreen(navController)
         }
         composable(
             "newItem/{edit}",
@@ -135,7 +117,6 @@ fun LoginApp(navController: NavController) {
 
     var popupMessage by remember { mutableStateOf<String?>(null) }
     var showPopup by remember { mutableStateOf(false) }
-    var popupColour by remember { mutableStateOf(colourError) }
 
     val isValidEmail = Patterns.EMAIL_ADDRESS.matcher(username).matches()
 
@@ -144,65 +125,28 @@ fun LoginApp(navController: NavController) {
         auth.signInWithEmailAndPassword(username, password)
             .addOnCompleteListener { task ->
                 if (task.isSuccessful) {
-                    val user = auth.currentUser
-                    if (user != null) {
-                        val uid = user.uid
-                        val userDocRef = db.collection("users").document(uid)
+                    val uid = auth.currentUser?.uid ?: return@addOnCompleteListener
+                    val userDocRef = db.collection("users").document(uid)
 
-                        userDocRef.get()
-                            .addOnSuccessListener { doc ->
-                                val userEmail = user.email ?: ""
-                                if (!doc.exists()) {
-                                    userDocRef.set(
-                                        mapOf(
-                                            "2faEnabled" to true,
-                                            "email" to userEmail,
-                                            "userType" to "unapproved",
-                                            "name" to "New User"
-                                        )
-                                    ).addOnSuccessListener {
-                                        navController.navigate("waiting") {
-                                            popUpTo("login") { inclusive = true }
-                                        }
-                                        isLoading = false
-                                    }
-                                } else {
-                                    if (doc.getBoolean("2faEnabled") == null) {
-                                        userDocRef.update("2faEnabled", true)
-                                    }
-
-                                    val type = doc.getString("userType") ?: "unapproved"
-                                    if (type == "unapproved") {
-                                        navController.navigate("waiting") {
-                                            popUpTo("login") { inclusive = true }
-                                        }
-                                        isLoading = false
-                                    } else {
-                                        navController.navigate("dashboard") {
-                                            popUpTo("login") { inclusive = true }
-                                        }
-                                        isLoading = false
-                                    }
-                                }
+                    userDocRef.get()
+                        .addOnSuccessListener { doc ->
+                            if (!doc.exists()) {
+                                userDocRef.set(mapOf("2faEnabled" to true))
+                            } else if (doc.getBoolean("2faEnabled") == null) {
+                                userDocRef.update("2faEnabled", true)
                             }
-                            .addOnFailureListener { e ->
-                                popupMessage = "Error: ${e.message}"
-                                showPopup = true
-                                isLoading = false
-                                popupColour = colourError
-                            }
-                    } else {
-                        popupMessage = "Login Failed"
-                        showPopup = true
-                        auth.signOut()
-                        isLoading = false
-                        popupColour = colourError
-                    }
+                            navController.navigate("dashboard")
+                            isLoading = false
+                        }
+                        .addOnFailureListener { e ->
+                            popupMessage = "Error: ${e.message}"
+                            showPopup = true
+                            isLoading = false
+                        }
                 } else {
                     popupMessage = task.exception?.message ?: "Login failed"
                     showPopup = true
                     isLoading = false
-                    popupColour = colourError
                 }
 
                 username = ""
@@ -266,7 +210,6 @@ fun LoginApp(navController: NavController) {
                         } else {
                             popupMessage = "Invalid email or password"
                             showPopup = true
-                            popupColour = colourError
                         }
                     }
                 ),
@@ -292,7 +235,6 @@ fun LoginApp(navController: NavController) {
                     } else {
                         popupMessage = "Invalid email or password"
                         showPopup = true
-                        popupColour = colourError
                     }
                 },
                 enabled = !isLoading,
@@ -308,9 +250,6 @@ fun LoginApp(navController: NavController) {
             ) {
                 Text("Login", fontSize = 16.sp, fontWeight = FontWeight.Bold)
             }
-            TextButton(onClick = { navController.navigate("signup") }) {
-                Text("Don't Have An Account? Sign Up", color = colourButton)
-            }
         }
 
         if (isLoading) {
@@ -334,196 +273,12 @@ fun LoginApp(navController: NavController) {
                 PopupMessage(
                     message = popupMessage!!,
                     onDismiss = { showPopup = false },
-                    backgroundColor = popupColour,
+                    backgroundColor = colourError,
                 )
             }
         }
     }
 }
 
-@Composable
-fun SignUpScreen(navController: NavController) {
-    var email by remember { mutableStateOf("") }
-    var password by remember { mutableStateOf("") }
-    var confirmPassword by remember { mutableStateOf("") }
-    var isLoading by remember { mutableStateOf(false) }
-    var popupMessage by remember { mutableStateOf<String?>(null) }
-    var popupColour by remember { mutableStateOf(colourError) }
-    var showPopup by remember { mutableStateOf(false) }
 
-    val isValidEmail = Patterns.EMAIL_ADDRESS.matcher(email).matches()
-    val scope = rememberCoroutineScope()
 
-    val signUp: () -> Unit = {
-        when {
-            !isValidEmail -> {
-                popupMessage = "Please enter a valid email."
-                showPopup = true
-                popupColour = colourError
-            }
-            password.isEmpty() || confirmPassword.isEmpty() -> {
-                popupMessage = "Password cannot be empty."
-                showPopup = true
-                popupColour = colourError
-            }
-            password != confirmPassword -> {
-                popupMessage = "Passwords do not match."
-                showPopup = true
-                popupColour = colourError
-            }
-            else -> {
-                isLoading = true
-                auth.createUserWithEmailAndPassword(email, password)
-                    .addOnCompleteListener { task ->
-                        isLoading = false
-                        if (task.isSuccessful) {
-                            popupMessage = "Account created! You can now log in."
-                            showPopup = true
-                            popupColour = colourButton
-                            auth.signOut()
-
-                            scope.launch {
-                                delay(2000)
-                                navController.navigate("waiting") {
-                                    popUpTo("login") { inclusive = true }
-                                }
-                            }
-                        } else {
-                            val exception = task.exception
-                            popupMessage = when {
-                                exception?.message?.contains("email address is already in use") == true ->
-                                    "An account with this email already exists."
-                                exception?.message?.contains("weak-password") == true ->
-                                    "Password is too weak. Please choose a stronger one."
-                                else -> exception?.message ?: "Sign up failed."
-                            }
-                            showPopup = true
-                            popupColour = colourError
-                        }
-                    }
-            }
-        }
-    }
-
-    Box(modifier = Modifier.fillMaxSize()) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(colourBackground),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Image(painter = painterResource(id = R.drawable.fearlessfalcons), contentDescription = "Logo")
-            Spacer(Modifier.height(10.dp))
-
-            Text("Sign Up", fontSize = 48.sp, fontWeight = FontWeight.Bold)
-            Spacer(modifier = Modifier.height(16.dp))
-
-            val customTextFieldColors = OutlinedTextFieldDefaults.colors(
-                focusedContainerColor = colourSecondary,
-                unfocusedContainerColor = colourSecondary,
-                focusedTextColor = colourSecondaryText,
-                unfocusedTextColor = colourSecondaryText,
-                focusedBorderColor = colourSecondaryText,
-                unfocusedBorderColor = colourSecondaryText,
-                cursorColor = colourButton,
-                selectionColors = TextSelectionColors(
-                    handleColor = colourButton,
-                    backgroundColor = colourButton.copy(alpha = 0.4f)
-                )
-            )
-
-            OutlinedTextField(
-                value = email,
-                onValueChange = { email = it },
-                label = { Text("Email") },
-                singleLine = true,
-                isError = email.isNotEmpty() && !isValidEmail,
-                modifier = Modifier
-                    .padding(8.dp)
-                    .width(300.dp),
-                textStyle = TextStyle(fontSize = 18.sp),
-                colors = customTextFieldColors,
-                keyboardOptions = KeyboardOptions(imeAction = ImeAction.Next)
-            )
-
-            OutlinedTextField(
-                value = password,
-                onValueChange = { password = it },
-                label = { Text("Password") },
-                singleLine = true,
-                visualTransformation = PasswordVisualTransformation(),
-                modifier = Modifier
-                    .padding(8.dp)
-                    .width(300.dp),
-                textStyle = TextStyle(fontSize = 18.sp),
-                colors = customTextFieldColors,
-                keyboardOptions = KeyboardOptions(imeAction = ImeAction.Next)
-            )
-
-            OutlinedTextField(
-                value = confirmPassword,
-                onValueChange = { confirmPassword = it },
-                label = { Text("Confirm Password") },
-                singleLine = true,
-                visualTransformation = PasswordVisualTransformation(),
-                modifier = Modifier
-                    .padding(8.dp)
-                    .width(300.dp),
-                textStyle = TextStyle(fontSize = 18.sp),
-                colors = customTextFieldColors,
-                keyboardOptions = KeyboardOptions(imeAction = ImeAction.Done)
-            )
-
-            Spacer(modifier = Modifier.height(16.dp))
-            FilledTonalButton(
-                shape = RoundedCornerShape(30),
-                onClick = { signUp() },
-                enabled = !isLoading,
-                colors = ButtonDefaults.filledTonalButtonColors(
-                    containerColor = colourButton,
-                    contentColor = colourBackground
-                ),
-                modifier = Modifier
-                    .padding(14.dp)
-                    .clip(RoundedCornerShape(2.dp))
-                    .fillMaxWidth()
-                    .height(56.dp),
-            ) {
-                Text("Create Account", fontSize = 16.sp, fontWeight = FontWeight.Bold)
-            }
-
-            Spacer(modifier = Modifier.height(8.dp))
-            TextButton(onClick = { navController.navigate("login") }) {
-                Text("Already have an account? Log in", color = colourButton)
-            }
-        }
-
-        if (isLoading) {
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .background(colourBackground.copy(alpha = 0.6f)),
-                contentAlignment = Alignment.Center
-            ) {
-                CircularProgressIndicator(color = colourButton)
-            }
-        }
-
-        if (showPopup && popupMessage != null) {
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(bottom = 64.dp),
-                contentAlignment = Alignment.BottomCenter
-            ) {
-                PopupMessage(
-                    message = popupMessage!!,
-                    onDismiss = { showPopup = false },
-                    backgroundColor = popupColour,
-                    textColor = colourBackground
-                )
-            }
-        }
-    }
-}
